move background devnet from util.rs to utils
