+++ crates/starknet-server/tests/get_transaction_receipt_by_hash.rs
@@ -18,8 +18,10 @@ mod get_transaction_receipt_by_hash_integration_tests {
     use starknet_rs_signers::{LocalWallet, SigningKey};
     use starknet_types::felt::Felt;
 
-    use crate::common::data::get_events_contract_in_sierra_and_compiled_class_hash;
-    use crate::common::util::{get_json_body, BackgroundDevnet};
+    use crate::common::devnet::BackgroundDevnet;
+    use crate::common::utils::{
+        get_events_contract_in_sierra_and_compiled_class_hash, get_json_body,
+    };
 
     #[tokio::test]
     async fn deploy_transaction_receipt() {
