+++ crates/starknet-server/tests/test_get_class_hash_at.rs
@@ -8,7 +8,7 @@ mod get_class_hash_at_integration_tests {
     };
 
     use crate::common::constants::PREDEPLOYED_ACCOUNT_ADDRESS;
-    use crate::common::util::BackgroundDevnet;
+    use crate::common::devnet::BackgroundDevnet;
 
     #[tokio::test]
     async fn get_class_hash_at_happy_path() {
