+++ crates/starknet-server/tests/test_minting.rs
@@ -7,7 +7,8 @@ mod minting_tests {
     use crate::common::constants::{
         PREDEPLOYED_ACCOUNT_ADDRESS, PREDEPLOYED_ACCOUNT_INITIAL_BALANCE,
     };
-    use crate::common::util::{get_json_body, BackgroundDevnet};
+    use crate::common::devnet::BackgroundDevnet;
+    use crate::common::utils::get_json_body;
 
     static DUMMY_ADDRESS: &str = "0x42";
     static DUMMY_AMOUNT: u128 = 42;
